Correct the syntax for file attachments
The sending of multiple files via thunderbird has been broken for a
while now. To reproduce, try to select multiple files in Nemo and select
send by mail.
The reason is, that the given syntax is not correct. Multiple file have
to be separated by only a colon, whereas the given separator also
includes a blank.

See also:
http://kb.mozillazine.org/Command_line_arguments_-_Thunderbird#Compose_new_mail_with_command_line

Removing the blank restores the functionality.