Disable auto-pinning of add-ons.
Explicitly installed runtimes are usually pinned automatically,
meaning they'll never be marked as 'unused' and ready for removal.

We know when an operation is for an addon, so disable this pinning
in these cases, so removing the parent app will also mark the
addons for removal. This is simpler than sifting thru addons
to check if they can be removed safely.

The method for this is missing from the libflatpak documentation,
so wasn't noticed sooner.