flatpaks: Prefer using a locally stored icon over a remote one, and return an icon of the correct size.
The icons defined in the appdata usually have any remote icons
after the local ones, but not always, and if the download fails
the app will end up with no icon at all.

Also, only the first size requested was being cached - this was
causing the package details page to use a smaller icon than
intended, and it was being scaled, making it not as sharp.