flatpaks: Query xmlb data for a developer name and cache it.
The 'developer' appstream field isn't supported by older appstream
versions - 'developer_name' is used.

Newer versions support it, and internally it also checks the
'developer_name' for backward compatibility.

Flathub apps are currently a mix of developer and developer_name
entries, so we use xmlb instead to check both, as well as the
'project_group' field, and cache it.

Caching it means we can remove the lazy updating of the xmlb data
when we later use the cache.