flatpak: Don't add installed runtimes to the in-memory package cache.
This was never intended and ends up confusing aspects of the
installer that assume every cached package has an appstream entry.

There are some standalone addon runtimes that *do* have appstream
entries - often multiple ones that apply to different platform
versions (branches). At this time distinguishing them seems to
be a challenge for the library we use to manage this, so they will
be treated like any other runtime.

Fixes linuxmint/mint22-beta#153.