pkgInfo: Filter out apt packages without a candidate, they're broken to mintinstall/mintcommon.
apt pkginfo: Check for a package candidate before accessing it.

This second change isn't currently necessary if we're filtering
these candidate-less packages now, but we're currently missing
runtime updating of un-cached (recently installed and/or foreign)
apt packages. We do this with flatpaks already, we should for debs
as well.