Fix Flatpak addon lookup.
An app's addons can be built with different runtime versions and
still remain compatible with a given version of the app.

What matters is the extension 'version' in the app's metadata,
and that it ends up matching the addon's FlatpakRef branch (which
is seemingly the only place to match the version string). Makes
perfect sense. As a bonus, an app can support multiple versions
of an extension point as well.