_flatpak.py: Don't try to pass a NULL branch to construct a FlatpakRemoteRef.
In libflatpak, even though FlatpakRef branch property defaults to
NULL, any attempt to set it assumes the value passed is non-NULL,
resulting in a segfault when performing a string search.

Assume 'master' if a broken .flatpakref file doesn't provide a Branch
field. This is a guess, but the only reasonable value to pass.

Fixes linuxmint/mintinstall#448.